% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_log.R
\name{analyse_log}
\alias{analyse_log}
\title{Analyse the Log}
\usage{
analyse_log(file)
}
\arguments{
\item{file}{path to the log file}
}
\value{
a tibble with one row per quiz
}
\description{
Read in and analyse the log file that is written by the WordBox shiny
app.
}
