% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_config}
\alias{get_config}
\title{Get Configuration Data}
\usage{
get_config(dir)
}
\arguments{
\item{dir}{character giving a directory path. If
a file called \code{wordbox.cfg} is found in that
folder it is read.}
}
\value{
a list containing the configuration data.
}
\description{
Read a configuration data from a file or use default values.
}
\details{
The function looks for a file called \code{wordbox.cfg}
first in the folder indicated by \code{dir} and, if the
file is not found, in the user's home. If no config file
is found in both locatings, default values are used.
}
