% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordlist.R
\name{get_filename}
\alias{get_filename}
\title{Get the File Name from which a Wordlist Was Read}
\usage{
get_filename(wl)
}
\arguments{
\item{wl}{a \code{wordlist} object}
}
\value{
a character with the path to the wordlist file
}
\description{
Get the File Name from which a Wordlist Was Read
}
