% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordlist.R
\name{get_groups}
\alias{get_groups}
\title{Extract the Names of All the Groups in a Wordlist}
\usage{
get_groups(wl)
}
\arguments{
\item{wl}{a \code{wordlist} object}
}
\value{
a character vector giving the names of the groups.
Each name is unique and the vector is sorted.
}
\description{
Extract the names of all the groups that are contained
in a \code{wordlist}.
}
