% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{get_languages}
\alias{get_languages}
\title{Get the Languages from a wordlist}
\usage{
get_languages(wl)
}
\arguments{
\item{wl}{a \code{wordlist} object}
}
\value{
a named vector with the names of the two languages
}
\description{
Extract the full names of the two languages that
are associated with a \code{wordlist}.
}
