% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordquiz.R
\name{get_quiz_cols}
\alias{get_quiz_cols}
\alias{get_quiz_type}
\title{Extract Attributes From a wordquiz Object}
\usage{
get_quiz_cols(quiz)

get_quiz_type(quiz)
}
\arguments{
\item{quiz}{a \code{\link{wordquiz}} object}
}
\description{
Extract the column names that are used for the quiz
(\code{quiz_cols}) and the quiz type (\code{quiz_type})
from a \code{wordquiz} object.
}
