% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_words.R
\name{mark_word}
\alias{mark_word}
\title{Register Success or Failure for a Word in a wordlist}
\usage{
mark_word(question, quiz, wl, success)
}
\description{
Register Success or Failure for a Word in a wordlist
}
