% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordquiz.R
\name{prepare_quiz}
\alias{prepare_quiz}
\title{Prepare an Quiz from a Wordlist}
\usage{
prepare_quiz(wl, direction)
}
\arguments{
\item{wl}{a \code{wordlist} object}

\item{direction}{an integer indicating the direction to be
 queried. \code{1} stands for Language1 to Language2 and
 \code{2} for the opposite direction.

 @return
 a \code{wordquiz} object}
}
\description{
Prepare an Quiz from a Wordlist
}
