% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{read_config}
\alias{read_config}
\title{Read a Configuration File}
\usage{
read_config(file)
}
\arguments{
\item{file}{character giving the path to the file}
}
\value{
a list containing the configuration data.
}
\description{
Read a configuration data from a file or use default values.
}
\details{
The configuration file must contain three lines
beginning with the keywords \code{BOXES},
\code{COUNTS} and \code{DAYS}.
}
\keyword{internal}
