% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{read_wordlist}
\alias{read_wordlist}
\title{Read a word list file}
\usage{
read_wordlist(file)
}
\arguments{
\item{file}{character indicating the name of the file}
}
\value{
a wordlist object, which is a data frame with the
following columns:
\describe{
\item{\code{language1}}{words in the first language}
\item{\code{language2}}{words in the second langauge}
\item{\code{group}}{name of the group that the words belong to}
\item{\code{box1}}{the box in which the words are in the mode 1 > 2}
\item{\code{count1}}{the number of consecutive successes in the
 current box in the mode 1 > 2}
\item{\code{date1}}{the date of the last success in the mode 1 > 2}
\item{\code{box2}}{the box in which the words are in the mode 2 > 1}
\item{\code{count2}}{the number of consecutive successes in the
 current box in the mode 2 > 1}
\item{\code{date2}}{the date of the last success in the mode 2 > 1}
}

The full names of the two languages are returned as attribut
and can be obtaines using \code{get_languages}.
}
\description{
Read a word list file
}
