% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{run_wordbox_dashboard}
\alias{run_wordbox_dashboard}
\title{Run Dashboard for Log and Wordlist Analysis}
\usage{
run_wordbox_dashboard(
  log_file = NULL,
  wordbox_dir = NULL,
  launch.browser = NULL
)
}
\arguments{
\item{log_file}{path to the log file}

\item{wordbox_dir}{path to the directory with the wordlist files}

\item{launch.browser}{logical, if \code{TRUE}, the application
is opened in the system's default browser, if \code{FALSE},
no browser is started. If the argument is omitted, the value
according to the option \code{shiny.launch.browser} is used,
which in RStudio opens the internal shiny viewer.}
}
\description{
Run Dashboard for Log and Wordlist Analysis
}
