% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_words.R
\name{update_quiz}
\alias{update_quiz}
\title{Update Quiz After Question Was Answered}
\usage{
update_quiz(question, quiz, wl, success)
}
\arguments{
\item{question}{the \code{\link{wordquestion}} object that
has been quizzed.}

\item{quiz}{the \code{\link{wordquiz}} from which the question
was taken and which will be modified.}

\item{wl}{the \code{\link{wordlist}}, on which the quiz is
based.}

\item{success}{logical indicating whether the question was
answered correctly or not.}
}
\description{
After a question from a quiz was answered, the
\code{wordquiz} object must be updated. This function
performs the update depending on the success of the answer
and the type of the quiz.
}
\details{
If an wrong answer was given, the \code{wordquiz} object
is returned unchanged. If the answer was correct, the
word that has been quizzed is removed from the quiz.

If the quiz type is \code{"newwords"}, a word is only
removed from the quiz once it has reached a count of 2.
Also, when a word is removed, the weights are adapted,
such that an additional word is included into the quiz.
}
