% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WordBox.R
\docType{package}
\name{WordBox-package}
\alias{WordBox}
\alias{WordBox-package}
\title{WordBox: a Simple Vocabulary Trainer}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
