% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordlist.R
\name{write_wordlist}
\alias{write_wordlist}
\title{Save a wordlist to a File}
\usage{
write_wordlist(wl, file, overwrite = FALSE)
}
\arguments{
\item{wl}{a \code{wordlist} object}

\item{file}{character giving the name of the file}

\item{overwrite}{logical indicating whether an existing file
should be overwritten}
}
\description{
Save a \code{wordlist} to a csv file
}
