% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordquiz.R
\name{correct_answer}
\alias{correct_answer}
\title{Check the Answer To a Question}
\usage{
correct_answer(answer, question)
}
\arguments{
\item{answer}{character string giving the answer to the question}

\item{question}{the \code{\link{wordquestion}} object that was
queried.}
}
\value{
a logical indicating whether the answer is correct
}
\description{
Check the Answer To a Question
}
