% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_wordlist.R
\name{plot_wordlist_counts}
\alias{plot_wordlist_counts}
\title{Plot a Wordlist}
\usage{
plot_wordlist_counts(
  wl,
  direction = 1,
  x = c("date", "box", "group"),
  colour = c("box", "group"),
  language = c("en", "de"),
  interactive = rlang::is_installed("plotly")
)
}
\arguments{
\item{wl}{a \code{wordlist} object}

\item{direction}{which direction to analyse. This can be given as an integer
(1: query left column; 2: query right column) or as a character string
indicating the query language as it is stored in the \code{wordlist}.}

\item{x}{the variable to be used for the x-axis}

\item{colour}{the variable used for colour}

\item{language}{language for the plot labels}

\item{interactive}{create an interactive plot? This requires `plotly` and
is set to `TRUE` by default, if `plotly` is installed. If set to `TRUE` by
the user, the function will ask to install `plotly` if no existing
installation is found.}
}
\value{
A `ggplot2` or `plotly` plot.
}
\description{
Plot a Wordlist
}
