% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordquiz.R
\name{prepare_quiz}
\alias{prepare_quiz}
\alias{wordquiz}
\title{Prepare an Quiz from a Wordlist}
\usage{
prepare_quiz(
  wl,
  direction,
  quiz_type = c("standard", "training", "newwords"),
  groups = NULL,
  n_words = Inf,
  core_only = FALSE,
  exam_only = FALSE,
  log_file = NULL
)
}
\arguments{
\item{wl}{a \code{wordlist} object}

\item{direction}{an integer indicating the direction to be
queried. \code{1} stands for Language1 to Language2 and
\code{2} for the opposite direction.}

\item{quiz_type}{character indicating the type of quiz to be run.
See 'details'.}

\item{groups}{character vector indicating the groups to
be quizzed. If omitted, all groups are included in the quiz.}

\item{n_words}{numeric value giving the maximum number of words
to be included in the quiz.}

\item{core_only}{logical. Should the quiz only involve words
that are part of the core vocabulary.}

\item{exam_only}{logical. Should the quiz only involve words
that are part of the current exam.}

\item{log_file}{If a valid path is provided, the quiz will write
its log to the file. If the file exists and is a log file
produced by WordBox, new entries will be appended. A file
that was not produced by WordBox will not be modified.}
}
\value{
a \code{wordquiz} object
}
\description{
Prepare an Quiz from a Wordlist
}
\details{
Three types of quizzes can be generated:
\describe{
  \item{standard}{Words from all boxes are included in the quiz,
   if the last correct answer is old enough. How old a correct
   answer must be for the word to be quizzed again depends on
   the box that it is in. The probability that a word is
   quizzed depends on the time that passed since the last
   correct answer and the box that it is in. Words in lower
   boxes or where the correct answer is further in the past
   are quizzed more often.}
  \item{training}{All words from all boxes are included in the
   quiz with equal probability weight.}
  \item{newwords}{Quiz only new words (words in box 1 with less than two
  correct answers). Questions are drawn from a list of three words only.
  A word is removed from this list and replaced after it has been answered
  correctly twice.}
}
The type of the quiz can be chosen by passing the appropriate
value to \code{quiz_type}.
}
