% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{read_config}
\alias{read_config}
\alias{get_default_config_file}
\title{Read a Configuration File}
\usage{
read_config(file)

get_default_config_file()
}
\arguments{
\item{file}{character giving the path to the file}
}
\value{
a list containing the configuration data.
}
\description{
Read WordBox configuration from a file.
}
\details{
The configuration file must be a JSON file containing a
single object with three fields:
\describe{
  \item{\code{"boxes"}}{a single integer indicating the number of
   boxes to be used.}
  \item{\code{"counts"}}{an integer array with one less elements
   than there are boxes. The numbers indicate for each box, how
   often the word must be correctly answered before it is moved
   to the next box. For the last box, the value is automatically
   set to \code{Inf}, since a word cannot be moved further than
   the last box.}
  \item{\code{"days"}}{an integer array with one element per box.
   For each box, this gives the number of days that a word is not
   quizzed, after it has been correctly answered.}
  \item{\code{"counts_new"}}{a single integer. Words in box 1 with
   count less than \code{counts_new} are considered to be new and
   are thus quizzed when \code{quiz_type = "newwords"} is used
   in \code{\link{prepare_quiz}}.}
  \item{\code{"n_new"}}{a single integer that defines how many
   words are quizzed at a time when \code{quiz_type = "newwords"}.}
}

\code{get_default_config_file()} returns the path to the
default configuration file. To look at its contents, run

\preformatted{
file.show(get_default_config_file())
}
}
\keyword{internal}
