% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordlist.R
\name{read_wordlist}
\alias{read_wordlist}
\alias{wordlist}
\alias{get_wordlist_testfile}
\title{Read a word list file}
\usage{
read_wordlist(file, config_file = NULL)

get_wordlist_testfile()
}
\arguments{
\item{file}{character with the full path to the word list file}

\item{config_file}{character giving the full path to the
configuration file. If omitted, the default configuration is
used. See \code{\link{read_config}} for details
on the config file.}
}
\value{
a wordlist object, which is a data frame with the
following columns:
\describe{
\item{\code{language1}}{words in the first language}
\item{\code{language2}}{words in the second language}
\item{\code{group}}{name of the group that the words belong to}
\item{\code{core}}{is the word part of the core vocabulary?}
\item{\code{exam}}{is the word part of the current exam?}
\item{\code{box1}}{the box in which the words are in the mode 1 > 2}
\item{\code{count1}}{the number of consecutive successes in the
 current box in the mode 1 > 2}
\item{\code{date1}}{the date of the last success in the mode 1 > 2}
\item{\code{box2}}{the box in which the words are in the mode 2 > 1}
\item{\code{count2}}{the number of consecutive successes in the
 current box in the mode 2 > 1}
\item{\code{date2}}{the date of the last success in the mode 2 > 1}
}

The full names of the two languages are returned as attribute of
the wordlist and can be obtained using
\code{\link{get_languages}}.
}
\description{
Read a word list file
}
\details{
\code{get_wordlist_testfile()} returns the file name of a wordlist
file that can be used for tests.
}
