% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{run_wordbox}
\alias{run_wordbox}
\title{Run the application}
\usage{
run_wordbox(dir = NULL)
}
\arguments{
\item{dir}{character indicating a directory where wordlist
files (csv) are stored. All the csv files in that directory
will be available to be loaded in the application.}
}
\description{
Run the application
}
