% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{run_wordbox}
\alias{run_wordbox}
\title{Run the application}
\usage{
run_wordbox(
  dir = NULL,
  launch.browser = NULL,
  config_file = NULL,
  log_file = NULL,
  show_errors = TRUE,
  n_words_default = 30,
  rm_trailing_chars = ""
)
}
\arguments{
\item{dir}{character indicating a directory where wordlist
files (csv) are stored. All the csv files in that directory
will be available to be loaded in the application.}

\item{launch.browser}{logical, if \code{TRUE}, the application
is opened in the system's default browser, if \code{FALSE},
no browser is started. If the argument is omitted, the value
according to the option \code{shiny.launch.browser} is used,
which in RStudio opens the internal shiny viewer.}

\item{config_file}{character giving the full path to the
configuration file. If omitted, the default configuration is
used. See \code{\link{read_config}} for details
on the config file.}

\item{log_file}{If a valid path is provided, the quiz will write
its log to the file. If the file exists and is a log file
produced by WordBox, new entries will be appended. A file
that was not produced by WordBox will not be modified.}

\item{show_errors}{should the UI show the number of errors
that were made during a quiz.}

\item{n_words_default}{integer indicating the default for the
number of words per quiz. This value must lie between 5
and 95.}

\item{rm_trailing_chars}{a character of length one containing
characters that should be removed if they appear at the
end of an answer. This can be used to omit rejecting answers
because a key next to Return is hit by accident. For a
Swiss keyboard, \code{rm_trailing_chars = "$"}
is an appropriate choice.}
}
\description{
Run the application
}
