% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WordBox.R
\docType{package}
\name{WordBox-package}
\alias{WordBox}
\alias{WordBox-package}
\title{WordBox: a Simple Vocabulary Trainer}
\description{
This package provides a simple vocabulary trainer
 in the form of a shiny app. Wordlists are provided as csv files.
}
\keyword{internal}
