% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WordBox.R
\docType{package}
\name{WordBox-package}
\alias{WordBox}
\alias{WordBox-package}
\title{WordBox: a Simple Vocabulary Trainer}
\description{
A simple vocabulary trainer in the form of a shiny app.
 The user can create his own word lists in the form of csv-files.
 The interface is in German.
}
\keyword{internal}
